requirements.txt
fastapi
uvicorn
python-dotenv
tavily-python
anthropic
sentence-transformers
python-whois
tldextract
docling
Pillow
python-multipart
requests
